--- src/compiler/typescript_to_rust/azle_generate/src/candid_type/services/class_decl/mod.rs
@@ -1,3 +1,0 @@
-mod errors;
-mod to_service;
-mod vm_value_conversions;
