+++ src/compiler/typescript_to_rust/azle_generate/src/candid_type/services/mod.rs
@@ -3,8 +3,8 @@ use cdk_framework::act::node::candid::Service;
 use crate::ts_ast::TsAst;
 use get_service_class_decls::GetFlattenedServiceClassDecls;
 
-mod class_decl;
 mod get_service_class_decls;
+mod to_service;
 
 impl TsAst {
     pub fn build_services(&self) -> Vec<Service> {
